fix(pci-instances): remove divider for the last item
ref: #TAPC-4967
