+++ packages/manager/apps/pci-instances/src/pages/instances/instance/dashboard/components/InstanceGeneralInfoBlock.component.tsx
@@ -16,10 +16,8 @@ import {
 import DashboardCardLayout from './DashboardCardLayout.component';
 import PriceLabel from '@/components/priceLabel/PriceLabel.component';
 import { useDashboard } from '../hooks/useDashboard';
-import {
-  DashboardTileBlock,
-  DashboardTileText,
-} from './DashboardTile.component';
+import { DashboardTileBlock } from './tile/DashboardTile.component';
+import { DashboardTileText } from './tile/DashboardTileText.component';
 import { useDashboardPolling } from '../hooks/useDashboardPolling';
 import { TaskStatus } from '@/pages/instances/task/TaskStatus.component';
 import { ActionsMenu } from '@/components/menu/ActionsMenu.component';
@@ -170,7 +168,7 @@ const InstanceGeneralInfoBlock: FC = () => {
         </DashboardTileBlock>
       )}
       {instance?.isDeleteEnabled && (
-        <DashboardTileBlock isLoading={isInstanceLoading} withoutDivider>
+        <DashboardTileBlock isLoading={isInstanceLoading}>
           <OsdsLink
             color={ODS_THEME_COLOR_INTENT.error}
             href={hrefDeleteInstance}
