+++ packages/manager/apps/pci-instances/src/pages/instances/instance/dashboard/components/tile/DashboardTile.component.tsx
@@ -2,19 +2,18 @@ import { FC, PropsWithChildren } from 'react';
 import { Divider, Text, TEXT_PRESET } from '@ovhcloud/ods-react';
 import { LoadingCell } from '@/pages/instances/datagrid/components/cell/LoadingCell.component';
 
-type TDashboardTileProps = PropsWithChildren<{
+export type TDashboardTileProps = PropsWithChildren<{
   isLoading?: boolean;
   label?: string;
-  withoutDivider?: boolean;
 }>;
 
 export const DashboardTileBlock: FC<TDashboardTileProps> = ({
   label,
   isLoading = false,
-  withoutDivider = false,
   children,
 }) => (
   <dl className="flex flex-col gap-1 m-0">
+    <Divider spacing="6" className="w-full" />
     <dt>
       {label && (
         <div className="flex items-center gap-2">
@@ -25,15 +24,5 @@ export const DashboardTileBlock: FC<TDashboardTileProps> = ({
     <dd className="m-0">
       <LoadingCell isLoading={isLoading}>{children}</LoadingCell>
     </dd>
-    {!withoutDivider && <Divider spacing="6" className="w-full" />}
   </dl>
-);
-
-export const DashboardTileText: FC<TDashboardTileProps> = ({
-  children,
-  ...props
-}) => (
-  <DashboardTileBlock {...props}>
-    <Text>{children}</Text>
-  </DashboardTileBlock>
 );
